Introduce shortener package
Extract the k8s resource name shortener used by Ingress/GatewayAPI to an
individual package. This package will be used to for implementing Hive
Job name shortening in the subsequent commits.
