+++ operator/pkg/model/translation/gateway-api/translator.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cilium/cilium/operator/pkg/model"
 	"github.com/cilium/cilium/operator/pkg/model/translation"
 	ciliumv2 "github.com/cilium/cilium/pkg/k8s/apis/cilium.io/v2"
+	"github.com/cilium/cilium/pkg/shortener"
 )
 
 var _ translation.Translator = (*gatewayAPITranslator)(nil)
@@ -125,11 +126,11 @@ func getService(resource *model.FullyQualifiedResource, allPorts []uint32, label
 		})
 	}
 
-	shortenName := model.Shorten(resource.Name)
+	shortenName := shortener.ShortenK8sResourceName(resource.Name)
 
 	return &corev1.Service{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      model.Shorten(ciliumGatewayPrefix + resource.Name),
+			Name:      shortener.ShortenK8sResourceName(ciliumGatewayPrefix + resource.Name),
 			Namespace: resource.Namespace,
 			Labels: mergeMap(map[string]string{
 				owningGatewayLabel: shortenName,
@@ -155,11 +156,11 @@ func getService(resource *model.FullyQualifiedResource, allPorts []uint32, label
 }
 
 func getEndpoints(resource model.FullyQualifiedResource, labels, annotations map[string]string) *corev1.Endpoints {
-	shortedName := model.Shorten(resource.Name)
+	shortedName := shortener.ShortenK8sResourceName(resource.Name)
 
 	return &corev1.Endpoints{
 		ObjectMeta: metav1.ObjectMeta{
-			Name:      model.Shorten(ciliumGatewayPrefix + resource.Name),
+			Name:      shortener.ShortenK8sResourceName(ciliumGatewayPrefix + resource.Name),
 			Namespace: resource.Namespace,
 			Labels: mergeMap(map[string]string{
 				owningGatewayLabel: shortedName,
@@ -208,7 +209,7 @@ func decorateCEC(cec *ciliumv2.CiliumEnvoyConfig, resource *model.FullyQualified
 		cec.Labels = make(map[string]string)
 	}
 	cec.Labels = mergeMap(cec.Labels, labels)
-	cec.Labels[gatewayNameLabel] = model.Shorten(resource.Name)
+	cec.Labels[gatewayNameLabel] = shortener.ShortenK8sResourceName(resource.Name)
 	cec.Annotations = mergeMap(cec.Annotations, annotations)
 
 	return nil
